+++ docs/examples/globus-storage/globus.py
@@ -1,3 +1,5 @@
+from typing import TYPE_CHECKING
+
 import os
 
 from globus_sdk import (
@@ -12,14 +14,17 @@ from globus_sdk.tokenstorage import SimpleJSONFileAdapter
 from machinable import Storage
 from pydantic import BaseModel, Field
 
+if TYPE_CHECKING:
+    from machinable import Interface
+
 
 class Globus(Storage):
     class Config(BaseModel):
-        client_id: str
-        local_endpoint_id: str
-        local_endpoint_directory: str
-        remote_endpoint_id: str
-        remote_endpoint_directory: str
+        client_id: str = Field("???")
+        remote_endpoint_id: str = Field("???")
+        local_endpoint_id: str = Field("???")
+        remote_endpoint_directory: str = Field("???")
+        local_endpoint_directory: str = "~/"
         auth_filepath: str = "~/.globus-tokens.json"
 
     def __init__(self, version=None):
@@ -28,6 +33,7 @@ class Globus(Storage):
         self._auth_file = None
         self._authorizer = None
         self._transfer_client = None
+        self.active_tasks = []
 
     @property
     def auth_client(self):
@@ -81,18 +87,59 @@ class Globus(Storage):
             self._transfer_client = TransferClient(authorizer=self.authorizer)
         return self._transfer_client
 
-    def commit(self, interface: "Interface") -> None:
-        ...
+    def commit(self, interface: "Interface") -> str:
+        try:
+            src = os.path.abspath(interface.local_directory())
+            if not os.path.exists(src):
+                raise RuntimeError("Interface must be committed before storage")
+
+            # This is not a strict requirement since client might allow access
+            # if os.path.normpath(src) != os.path.normpath(self.local_path(interface.uuid)):
+            #     raise RuntimeError("Interface directory must be in storage directory")
+
+            task_data = TransferData(
+                source_endpoint=self.config.local_endpoint_id,
+                destination_endpoint=self.config.remote_endpoint_id,
+                notify_on_succeeded=False,
+                notify_on_failed=False,
+            )
+
+            task_data.add_item(
+                src,
+                self.remote_path(interface.uuid),
+                recursive=True,
+            )
+
+            task_doc = self.transfer_client.submit_transfer(task_data)
+
+            task_id = task_doc["task_id"]
+
+            self.active_tasks.append(task_id)
+
+            print(f"Submitted Globus commit, task_id={task_id}")
+
+            # this operation is non-blocking by default
+
+            return task_id
+        except TransferAPIError as e:
+            if e.code == "Conflict":
+                return False
+            elif e.code == "ConsentRequired":
+                raise RuntimeError(
+                    f"You do not have the right permissions. Try removing {self.config.auth_filepath} and authenticating again with the appropriate identity provider."
+                ) from e
+            raise e
 
     def update(self, interface: "Interface") -> None:
-        ...
+        return self.commit(interface)
 
     def contains(self, uuid: str) -> bool:
         # check if folder exists on globus storage
         try:
             response = self.transfer_client.operation_ls(
                 self.config.remote_endpoint_id,
-                path=os.path.join(self.config.remote_endpoint_directory, uuid),
+                path=self.remote_path(uuid),
+                show_hidden=True,
             )
         except TransferAPIError as e:
             if e.code == "ClientError.NotFound":
@@ -109,17 +156,41 @@ class Globus(Storage):
 
         return False
 
-    def retrieve(self, uuid: str, local_directory: str) -> bool:
-        # task_data = TransferData(
-        #     source_endpoint=self.config.remote_endpoint_id,
-        #     destination_endpoint=self.config.local_endpoint_id,
-        # )
-        # task_data.add_item(
-        #     os.path.join(self.config.remote_endpoint_directory, uuid),
-        #     target_directory,
-        # )
+    def retrieve(
+        self, uuid: str, local_directory: str, timeout: int = 5 * 60
+    ) -> bool:
+        if not self.contains(uuid):
+            return False
 
-        # task_doc = self.transfer_client.submit_transfer(task_data)
-        # task_id = task_doc["task_id"]
-        # print(f"submitted transfer, task_id={task_id}")
-        return False
+        task_data = TransferData(
+            source_endpoint=self.config.remote_endpoint_id,
+            destination_endpoint=self.config.local_endpoint_id,
+            notify_on_succeeded=False,
+            notify_on_failed=False,
+        )
+        task_data.add_item(
+            self.remote_path(uuid),
+            local_directory,
+        )
+        task_doc = self.transfer_client.submit_transfer(task_data)
+        task_id = task_doc["task_id"]
+        self.active_tasks.append(task_id)
+
+        print(f"[Storage] Submitted Globus retrieve, task_id={task_id}")
+
+        self.tasks_wait(timeout=timeout)
+
+        return True
+
+    def tasks_wait(self, timeout: int = 5 * 60) -> None:
+        for task_id in self.active_tasks:
+            print(f"[Storage] Waiting for Globus task {task_id} to complete")
+            self.transfer_client.task_wait(task_id, timeout=timeout)
+            print(f"[Storage] task_id={task_id} transfer finished")
+        self.active_tasks = []
+
+    def local_path(self, *append):
+        return os.path.join(self.config.local_endpoint_directory, *append)
+
+    def remote_path(self, *append):
+        return os.path.join(self.config.remote_endpoint_directory, *append)